Reorg MIR rewrite passes, and add debugging info
