+++ mir-verifier.cabal
@@ -40,5 +40,10 @@ library
 
   other-modules: 
                  Mir.Pass,
+                 Mir.Pass.CollapseRefs,
+                 Mir.Pass.MutRefReturnStatic,
+                 Mir.Pass.RewriteMutRef,
+                 Mir.Pass.RemoveStorage,
+                 Mir.Pass.RemoveBoxNullary,
                  Mir.Run,
                  Mir.Trans
