+++ src/Mir/SAWInterface.hs
@@ -3,6 +3,10 @@ module Mir.SAWInterface (RustModule, loadMIR, extractMIR, rmCFGs) where
 import Mir.Run
 import Mir.Mir
 import Mir.Pass as P
+import Mir.Pass.CollapseRefs as P
+import Mir.Pass.RewriteMutRef as P
+import Mir.Pass.RemoveStorage as P
+import Mir.Pass.RemoveBoxNullary as P
 import System.IO
 import System.FilePath
 import qualified Data.Text as T
@@ -30,6 +34,9 @@ import qualified Text.Regex as Regex
 
 import Control.Monad
 
+import GHC.Stack
+
+
 data RustModule = RustModule {
     rmCFGs :: M.Map T.Text C.AnyCFG
 }
@@ -52,13 +59,14 @@ extractMIR sc rm n = do
     term <- extractFromCFGPure link sc cfg
     return term
 
-loadMIR :: SC.SharedContext -> FilePath -> IO RustModule
+loadMIR :: HasCallStack => SC.SharedContext -> FilePath -> IO RustModule
 loadMIR sc fp = do
     f <- B.readFile fp
     let c = (J.eitherDecode f) :: Either String Collection
     case c of
       Left msg -> fail $ "Decoding of MIR failed: " ++ msg
       Right coll -> do
+          mapM_ (putStrLn . pprint) coll
           let cfgmap_ = mirToCFG coll (Just (P.passMutRefArgs . P.passRemoveStorage . P.passRemoveBoxNullary))
           let cfgmap = M.fromList $ map (\(k,v) -> (cleanFnName k, v)) $ M.toList cfgmap_
           return $ RustModule cfgmap
